feat: actually do the processing of template stuff inside the async run evaluation, remove the whole of the old non-async one (deprecated and was really confusing)
