+++ apps/experiments/src/langsmith/service.ts
@@ -1,18 +1,8 @@
-import {
-  createRitaGraph,
-  RitaThreadStatus,
-  RitaThreadTriggerType,
-} from "@the-project-b/rita-graphs";
 import { createLogger } from "@the-project-b/logging";
+import { createRitaGraph } from "@the-project-b/rita-graphs";
 import { Client } from "langsmith";
-import { evaluate } from "langsmith/evaluation";
 import { getAuthUser } from "../security/auth.js";
-import { createGraphQLClient } from "../graphql/index.js";
 
-import { GraphQLErrors } from "../graphql/errors.js";
-import { TemplateProcessor } from "../evaluators/helpers/template-processor.js";
-import { TemplateContext } from "../evaluators/helpers/template-variable-registry.js";
-import type { GraphQLContext } from "../types/context.js";
 import type {
   DatasetExperiment,
   DeleteExperimentRunsInput,
@@ -23,7 +13,6 @@ import type {
   GetExperimentDetailsInput,
   GraphName,
   Run,
-  RunEvaluationInput,
 } from "../types/index.js";
 
 // Define types for prompt information
@@ -86,282 +75,6 @@ export class LangSmithService {
    */
   public getAvailableGraphs(): GraphName[] {
     return Object.keys(this.graphFactoryMap) as GraphName[];
-  }
-
-  public async runEvaluation(
-    input: RunEvaluationInput,
-    context: GraphQLContext,
-  ) {
-    const {
-      graphName,
-      datasetName,
-      splits,
-      evaluators,
-      experimentPrefix,
-      selectedCompanyId,
-      maxConcurrency,
-      numRepetitions,
-    } = input;
-
-    if (!context.user) {
-      throw GraphQLErrors.UNAUTHENTICATED;
-    }
-
-    const graphFactory = this.graphFactoryMap[graphName];
-    if (!graphFactory) {
-      throw new Error(`Graph factory not found for graph: ${graphName}`);
-    }
-
-    // Create Rita thread in database BEFORE running evaluations
-    const graphqlClient = createGraphQLClient(context.token || "");
-
-    // Generate the LangGraph thread ID that will be used for all evaluations
-    const lcThreadId = `eval-${Date.now()}-${Math.random().toString(36).substring(7)}`;
-
-    const threadResult = await graphqlClient.createRitaThread({
-      input: {
-        triggerType: RitaThreadTriggerType.Evaluation,
-        hrCompanyId: selectedCompanyId,
-        status: RitaThreadStatus.Received,
-        lcThreadId, // Use the same thread ID for LangGraph
-      },
-    });
-
-    const ritaThread = threadResult.createRitaThread;
-    logger.info(
-      `🧵 RitaThread created: ${ritaThread.id} (lc: ${ritaThread.lcThreadId})`,
-      {
-        threadId: ritaThread.id,
-        lcThreadId: ritaThread.lcThreadId,
-        operation: "createRitaThread",
-        graphName,
-        datasetName,
-        experimentPrefix: experimentPrefix || graphName,
-        selectedCompanyId,
-        triggerType: RitaThreadTriggerType.Evaluation,
-      },
-    );
-
-    const target = async (inputs: Record<string, any>) => {
-      let question = inputs.question;
-
-      const templateContext: TemplateContext = {
-        currentDate: new Date(),
-      };
-
-      if (question && typeof question === "string") {
-        const templateResult = TemplateProcessor.process(
-          question,
-          templateContext,
-        );
-        question = templateResult.processed;
-
-        if (templateResult.replacements.length > 0) {
-          logger.debug("Template variables processed in question", {
-            operation: "processTemplates",
-            original: templateResult.original,
-            processed: templateResult.processed,
-            replacements: templateResult.replacements.map((r) => ({
-              expression: r.expression,
-              value: r.result.displayValue,
-            })),
-          });
-        }
-      }
-
-      const examplePreferredLanguage = inputs.preferredLanguage;
-
-      const graphInput = {
-        messages: [{ role: "user", content: question }],
-        ...(examplePreferredLanguage && {
-          preferredLanguage: examplePreferredLanguage,
-        }),
-      };
-      const graph = await graphFactory();
-
-      let token = context.token || "";
-      if (token.toLowerCase().startsWith("bearer ")) {
-        token = token.slice(7).trim();
-      }
-
-      const config = {
-        configurable: {
-          thread_id: ritaThread.lcThreadId, // Use the LangGraph thread ID for all evaluations
-          langgraph_auth_user: {
-            token,
-            user: {
-              firstName: context.user.me.firstName,
-              lastName: context.user.me.lastName,
-              preferredLanguage:
-                examplePreferredLanguage || context.user.me.preferredLanguage,
-              company: {
-                id: selectedCompanyId,
-              },
-            },
-          },
-        },
-      };
-
-      const result: any = await graph.invoke(graphInput, config);
-
-      const lastMessage = Array.isArray(result?.messages)
-        ? result.messages[result.messages.length - 1]
-        : undefined;
-      const answer = lastMessage?.content;
-
-      if (typeof answer !== "string") {
-        logger.warn(
-          `[${ritaThread.id}] Graph did not return a final message with string content`,
-          {
-            threadId: ritaThread.id,
-            lcThreadId: ritaThread.lcThreadId,
-            operation: "evaluateTarget",
-            graphName,
-            messageType: typeof answer,
-            hasMessages: Array.isArray(result?.messages),
-            messageCount: result?.messages?.length || 0,
-          },
-        );
-        return {
-          answer: "",
-          dataChangeProposals: [],
-        };
-      }
-
-      // Fetch data change proposals from the database after graph execution
-      const threadItemsResult = await graphqlClient.getThreadItemsByThreadId({
-        threadId: ritaThread.id,
-      });
-
-      const dataChangeProposals = [];
-      if (threadItemsResult?.thread?.threadItems) {
-        for (const item of threadItemsResult.thread.threadItems) {
-          try {
-            const data =
-              typeof item.data === "string" ? JSON.parse(item.data) : item.data;
-            if (data.type === "DATA_CHANGE_PROPOSAL" && data.proposal) {
-              // We want the actual proposal to be passed and the evaluator handles the normalization
-              logger.info(" ADDED proposaldata.proposal", data.proposal);
-              dataChangeProposals.push(data.proposal);
-            }
-          } catch (e) {
-            // Skip invalid JSON items
-          }
-        }
-      }
-
-      return {
-        answer,
-        dataChangeProposals,
-        threadTitle: threadItemsResult?.thread?.title || null,
-      };
-    };
-
-    // Prepare evaluators, potentially fetching prompts from LangSmith
-    const { createEvaluator } = await import("../evaluators/core/factory.js");
-
-    const evaluatorPromises = evaluators.map(async (evaluatorInput) => {
-      let promptToUse = evaluatorInput.customPrompt;
-
-      if (evaluatorInput.langsmithPromptName && !evaluatorInput.customPrompt) {
-        try {
-          const promptData = await this.pullPrompt(
-            evaluatorInput.langsmithPromptName,
-          );
-          promptToUse = this.convertPromptToText(promptData.promptData);
-        } catch (error) {
-          logger.error(
-            `Failed to fetch prompt ${evaluatorInput.langsmithPromptName}:`,
-            error,
-            {
-              operation: "fetchLangsmithPrompt",
-              promptName: evaluatorInput.langsmithPromptName,
-              evaluatorType: evaluatorInput.type,
-              hasCustomPrompt: !!evaluatorInput.customPrompt,
-              errorType:
-                error instanceof Error
-                  ? error.constructor.name
-                  : "UnknownError",
-              errorMessage:
-                error instanceof Error ? error.message : String(error),
-            },
-          );
-          throw new Error(
-            `Failed to fetch LangSmith prompt "${evaluatorInput.langsmithPromptName}": ${error instanceof Error ? error.message : "Unknown error"}`,
-          );
-        }
-      }
-
-      return createEvaluator(
-        evaluatorInput.type,
-        promptToUse,
-        evaluatorInput.model,
-        evaluatorInput.referenceKey,
-      );
-    });
-
-    const evaluationConfig = {
-      evaluators: await Promise.all(evaluatorPromises),
-      experimentPrefix: experimentPrefix || `eval-${graphName}`,
-      // Set concurrency for concurrent processing of dataset examples
-      maxConcurrency: maxConcurrency || 10, // Increased default for better concurrency
-      numRepetitions: numRepetitions || 1, // Number of times to run each example
-    };
-
-    logger.info("Running evaluation with splits ", {
-      datasetName,
-      splits,
-    });
-
-    // We filter based on splits here
-    const exampleGenerator = this.client.listExamples({
-      datasetName,
-      splits,
-    });
-
-    const experimentResults = await evaluate(target, {
-      ...evaluationConfig,
-      data: exampleGenerator,
-    });
-
-    const results = [];
-    for (const item of experimentResults.results ?? []) {
-      const run = item.run;
-      const evalResults = item.evaluationResults;
-      if (!run) continue;
-
-      const scores = (evalResults?.results ?? []).map((score) => ({
-        key: score.key,
-        score: String(score.score),
-        comment: score.comment,
-      }));
-
-      const startTime = Number(run.start_time ?? 0);
-      const endTime = Number(run.end_time ?? 0);
-      const latency = endTime > 0 && startTime > 0 ? endTime - startTime : 0;
-      const totalTokens = run.total_tokens ?? 0;
-
-      results.push({
-        id: run.id,
-        inputs: JSON.stringify(run.inputs),
-        outputs: run.outputs ? JSON.stringify(run.outputs) : null,
-        startTime: new Date(startTime).toISOString(),
-        endTime: new Date(endTime).toISOString(),
-        latency,
-        totalTokens,
-        scores,
-      });
-    }
-
-    const experimentName = experimentResults.experimentName;
-    const experimentId = experimentResults.results[0].run.id;
-
-    return {
-      url: "deprecated - please see langsmith dashboard directly",
-      experimentName,
-      experimentId,
-      results,
-    };
   }
 
   /**
